Add common CPU brands next to architectures
Now that we are soon getting arm64 support, one common concern shared on the launcher side was people not understanding if their system was x64 or arm64.

Is this overboard? Probably, but it should clear confusion.

Hopefully intel or amd does not start making arm based processors soon. To my knowledge, only snapdragon is the only other brand selling arm64 based cpu's for computers at this current time, but to be sure i added "etc" because pretty much anything else will probably be arm64 in the coming years of technology.